requirements.txt
adbutils==0.4.4
appdirs==1.4.3
Appium-Python-Client==0.47
APScheduler==3.6.1
arrow==0.15.0
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.0
blinker==1.4
bounded-pool-executor==0.0.3
brotlipy==0.7.0
bs4==0.0.1
cacheout==0.11.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
conda==4.3.16
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
dateparser==0.7.1
decorator==4.4.0
demjson==2.2.4
Deprecated==1.2.6
deprecation==2.0.7
docopt==0.6.2
et-xmlfile==1.0.1
facebook-wda==0.4.0
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==5.2.1
fire==0.2.1
Flask==1.1.1
goose3==3.1.6
grpcio==1.23.0
h11==0.9.0
h2==3.1.1
hpack==3.0.0
humanize==0.5.1
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itchat==1.3.10
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.39
jieba3k==0.35.1
Jinja2==2.10.1
js2xml==0.3.1
jsonpath==0.82
kaitaistruct==0.8
kiwisolver==1.1.0
ldap3==2.6.1
logzero==1.5.0
lxml==4.3.4
MarkupSafe==1.1.1
matplotlib==3.1.1
mitmproxy==4.0.4
moment==0.8.2
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.17.2
openpyxl==2.6.3
packaging==19.1
pandas==0.25.1
parse==1.12.1
parsel==1.5.2
passlib==1.7.1
pika==1.1.0
Pillow==6.1.0
pinyin==0.4.0
ply==3.11
progress==1.5
protobuf==3.9.1
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pydash==4.7.5
PyDispatcher==2.0.5
pyee==6.0.0
PyExecJS==1.5.1
PyHamcrest==1.9.0
pymongo==3.8.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyperclip==1.7.0
pypng==0.0.20
pyppeteer==0.0.25
PyQRCode==1.2.1
pyquery==1.4.0
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
queuelib==1.5.0
raven==6.10.0
readability-lxml==0.7.1
redis==3.3.8
regex==2019.8.19
requests==2.22.0
requests-file==1.4.3
requests-html==0.10.0
retry==0.9.2
retrying==1.3.3
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
Scrapy==1.7.3
scrapy-redis==0.6.8
scrapy-splash==0.7.2
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
slimit==0.8.1
sortedcontainers==2.1.0
soupsieve==1.9.3
termcolor==1.1.0
times==0.7
tinycss==0.4
tinysegmenter==0.4
tld==0.9.3
tldextract==2.2.1
tornado==6.0.3
tqdm==4.35.0
Twisted==19.7.0
tzlocal==2.0.0
uiautomator2==1.2.4
urllib3==1.25.3
urwid==2.0.1
w3lib==1.20.0
websockets==8.0.2
weditor==0.2.5
Werkzeug==0.15.6
whichcraft==0.6.1
wordcloud==1.5.0
wrapt==1.11.2
wsproto==0.15.0
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
you-get==0.4.1347
zope.interface==4.6.0
